+++ weblate/accounts/tests/test_registration.py
@@ -49,7 +49,7 @@ class RegistrationTest(TestCase, RegistrationTestMixin):
     def assert_registration(self, match=None):
         url = self.assert_registration_mailbox(match)
 
-        if self.clear_cookie:
+        if self.clear_cookie and 'sessionid' in self.client.cookies:
             del self.client.cookies['sessionid']
 
         # Confirm account
@@ -118,6 +118,46 @@ class RegistrationTest(TestCase, RegistrationTestMixin):
         # Verify stored first/last name
         self.assertEqual(user.first_name, 'First Last')
 
+    @OverrideSettings(REGISTRATION_OPEN=True)
+    @OverrideSettings(REGISTRATION_CAPTCHA=False)
+    def test_double_register(self):
+        """Test double registration from single browser"""
+
+        # First registration
+        response = self.client.post(
+            reverse('register'),
+            REGISTRATION_DATA
+        )
+        first_url = self.assert_registration_mailbox()
+        mail.outbox.pop()
+
+        # Second registration
+        data = REGISTRATION_DATA.copy()
+        data['email'] = 'noreply@example.net'
+        data['username'] = 'second'
+        response = self.client.post(
+            reverse('register'),
+            data,
+        )
+        second_url = self.assert_registration_mailbox()
+        mail.outbox.pop()
+
+        # Confirm first account
+        response = self.client.get(first_url, follow=True)
+        self.assertRedirects(
+            response,
+            reverse('password')
+        )
+        self.client.get(reverse('logout'))
+
+        # Confirm second account
+        response = self.client.get(second_url, follow=True)
+        self.assertRedirects(
+            response,
+            reverse('password')
+        )
+
+    @OverrideSettings(REGISTRATION_OPEN=True)
     @OverrideSettings(REGISTRATION_CAPTCHA=False)
     def test_register_missing(self):
         # Disable captcha
@@ -133,9 +173,6 @@ class RegistrationTest(TestCase, RegistrationTestMixin):
 
         # Remove session ID from URL
         url = url.split('&id=')[0]
-
-        # Delete session ID from cookies
-        del self.client.cookies['sessionid']
 
         # Confirm account
         response = self.client.get(url, follow=True)