+++ weblate/appsettings.py
@@ -222,15 +222,15 @@ HIDE_REPO_CREDENTIALS = getvalue('HIDE_REPO_CREDENTIALS', True)
 # GitHub username for sending pull requests
 GITHUB_USERNAME = getvalue('GITHUB_USERNAME', None)
 
-# Whiteboard
-ENABLE_WHITEBOARD = getvalue('ENABLE_WHITEBOARD', False)
-
 # Default committer
 DEFAULT_COMMITER_EMAIL = getvalue(
     'DEFAULT_COMMITER_EMAIL', 'noreply@weblate.org'
 )
 DEFAULT_COMMITER_NAME = getvalue(
     'DEFAULT_COMMITER_NAME', 'Weblate'
+)
+DEFAULT_TRANSLATION_PROPAGATION = getvalue(
+    'DEFAULT_TRANSLATION_PROPAGATION', True
 )
 
 # Obsolete configs, needed for data migration