+++ weblate/trans/models/whiteboard.py
@@ -21,11 +21,70 @@
 """Whiteboard model."""
 
 from django.db import models
-from django.utils.translation import ugettext_lazy
+from django.db.models import Q
+from django.utils.translation import ugettext_lazy, ugettext as _
+from django.core.exceptions import ValidationError
+from weblate.lang.models import Language
+
+
+class WhiteboardManager(models.Manager):
+    def context_filter(self, project=None, subproject=None, language=None):
+        """Filters whiteboard messages by context."""
+        base = self.all()
+
+        if language is None and project is None and subproject is None:
+            return base.filter(
+                project=None, subproject=None, language=None
+            )
+
+        if language and project is None and subproject is None:
+            return base.filter(
+                project=None, subproject=None, language=language
+            )
+
+        if project:
+            return base.filter(Q(project=project) & Q(subproject=None))
+
+        if subproject:
+            if language:
+                return base.filter(
+                    Q(language=language) |
+                    Q(subproject=subproject) |
+                    Q(project=subproject.project)
+                )
+
+            return base.filter(
+                Q(subproject=subproject) | Q(project=subproject.project)
+            )
+
+        return base
 
 
 class WhiteboardMessage(models.Model):
-    message = models.TextField(blank=True)
+    message = models.TextField(
+        verbose_name=ugettext_lazy('Message'),
+    )
+
+    project = models.ForeignKey(
+        'Project',
+        verbose_name=ugettext_lazy('Project'),
+        null=True,
+        blank=True,
+    )
+    subproject = models.ForeignKey(
+        'SubProject',
+        verbose_name=ugettext_lazy('Component'),
+        null=True,
+        blank=True
+    )
+    language = models.ForeignKey(
+        Language,
+        verbose_name=ugettext_lazy('Language'),
+        null=True,
+        blank=True
+    )
+
+    objects = WhiteboardManager()
 
     class Meta(object):
         app_label = 'trans'
@@ -34,3 +93,12 @@ class WhiteboardMessage(models.Model):
 
     def __unicode__(self):
         return self.message
+
+    def clean(self):
+        if self.project and self.subproject:
+            if self.subproject.project != self.project:
+                raise ValidationError(
+                    _('Do not specify both component and project!')
+                )
+        if not self.project and self.subproject:
+            self.project = self.subproject.project