+++ weblate/trans/views/basic.py
@@ -284,7 +284,7 @@ def show_translation(request, project, subproject, lang):
     obj.is_locked(request.user)
 
     # Get form
-    form = get_upload_form(request)()
+    form = get_upload_form(request.user, obj.subproject.project)()
 
     # Is user allowed to do automatic translation?
     if can_automatic_translation(request.user, obj.subproject.project):